pyproject.toml
[project]
name = "mcp-ibm-i"
version = "0.1.0"
description = "Secure MCP Gateway for IBM i (AS/400) systems with command validation and SSH connectivity"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Santiago Pernia" }
]
license = { text = "MIT" }
keywords = ["mcp", "ibm-i", "as400", "ssh", "gateway", "security"]
dependencies = [
    "mcp[cli]>=1.25.0",
    "paramiko>=4.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
ibmi-gateway = "ibmi_gateway.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ibmi_gateway"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

